Fix issue #12 - netnsinit fails if gateway not set
Closes https://github.com/Jamesits/systemd-named-netns/issues/12

"set -e" does not play well with shorthand "if" checks "[ a ] && b"
inside functions due to how exit code propagates to function exit code.
No such issue when using normal if-then because in such case exit code
is not set to 1 when check fails.
Additionally added "return 0" just in case future update to functions
add more potentially failty code - this is enough to prevent it from
failing. Not needed in current version though.